requirements.txt
python-telegram-bot==20.3
requests
python-dotenv
aiohttp
loguru
validators
beautifulsoup4
lxml
python-multipart
flask
gunicorn
